# e-commerce_backend/openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/sellers/{id}/products": {
            "get": {
                "operationId": "SellerController_getSellerWithProducts",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sellers"
                ]
            },
            "post": {
                "operationId": "SellerController_createProductForSeller",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sellers"
                ]
            }
        },
...
        "/api/v1/products/create-with-image": {
            "post": {
                "operationId": "ProductController_createProductWithImage",
                "summary": "Create product with image upload",
                "description": "Creates a new product with image upload. Requires ADMIN or SELLER role. Automatically generates image URL.",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Product creation data with image file",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Latest iPhone with A17 Pro chip"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 999.99
                                    },
                                    "stockQuantity": {
                                        "type": "number",
                                        "example": 50
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Electronics"
                                    },
                                    "isActive": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/products/paginated": {
            "get": {
                "operationId": "ProductController_getPaginatedProducts",
                "parameters": [
                    {
                        "name": "page",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "operationId": "ProductController_getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            },
            "put": {
                "operationId": "ProductController_updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            },
            "delete": {
                "operationId": "ProductController_deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            }
        },
        "/api/v1/products/create": {
            "post": {
                "operationId": "ProductController_createProduct",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            }
        },
...
        "/api/v1/products/seller/{sellerId}": {
            "get": {
                "operationId": "ProductController_getProductsBySellerId",
                "parameters": [
                    {
                        "name": "sellerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Products"
                ]
            }
        },
...
    },
    "info": {
        "title": "E-Commerce Platform API",
        "description": "Complete API documentation for the E-Commerce platform with authentication, products, orders, notifications, and seller management.",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:4002",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:4002/api/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Enter JWT token"
            },
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "JWT token stored in HTTP-only cookie"
            }
        },
        "schemas": {
...
            "CreateProductDto": {
                "type": "object",
                "properties": {}
            },
            "Product": {
                "type": "object",
                "properties": {}
            },
            "UpdateProductDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}